+++ test-unix.sh
@@ -2,30 +2,29 @@
 
 set -e
 
-PARANOID_FLAGS="-DGC_STRESS -DCHECK_OBJ_POINTERS -DTEST -DPARANOID -DGCSTACK_OVERFLOWCHECKS -DGCSTACK_UNDERFLOWCHECKS -DTAGSTACK_OVERFLOWCHECKS -DTAGSTACK_UNDERFLOWCHECKS"
+PARANOID="-DCHECK_OBJ_POINTERS -DTEST -DPARANOID -DGCSTACK_OVERFLOWCHECKS -DGCSTACK_UNDERFLOWCHECKS -DTAGSTACK_OVERFLOWCHECKS -DTAGSTACK_UNDERFLOWCHECKS"
 
-build_and_run () {
-    make clean all TARGET=unix LISP_FLAGS="$1" $2
-    cd tunix/unix/ && (echo | ./lisp) ; cd -
+test_sim6502 () {
+    make clean world TARGET=sim6502 LISP_FLAGS="$1" $2
+    cd tunix/sim6502/ && (echo "(isave \"image\")" | sim65 lisp) ; cd -
+    cd tunix/sim6502/ && (echo | sim65 lisp) ; cd -
 }
 
-test_options () {
-    build_and_run "$1" $2
-    build_and_run "-DCOMPRESSED_CONS $1" $2
-    build_and_run "-DNO_ONERROR $1" $2
-    build_and_run "-DNO_DEBUGGER $1" $2
-    build_and_run "-DNO_IMAGES $1" $2
+test_unix () {
+    make clean world TARGET=unix LISP_FLAGS="$1" $2
+    cd tunix/unix/ && (echo -n "(isave \"image\")(iload \"image\")" | ./lisp) ; cd -
 }
 
-stress_out () {
-    test_options "$PARANOID_FLAGS" "GC_STRESS=1"
-}
+# TODO: -DCHECK_OBJ_POINTERS on sim65.
+test_sim6502 "-DNDEBUG"
+test_sim6502 "-DCOMPRESSED_CONS"
+test_sim6502 "-DNO_ONERROR"
+test_sim6502 "-DNO_DEBUGGER"
+test_sim6502 "-DNAIVE"
 
-build_and_run "-DNAIVE"
-
-build_and_run "$PARANOID_FLAGS"
-cd tunix/unix/ && (echo "(isave \"image\")\n(iload \"image\")\n" | ./lisp) ; cd -
-# Not loadable.  Needs relocating.
-#cd tunix/unix/ && (echo | ./lisp) ; cd -
-
-#stress_out
+test_unix "$PARANOID"
+test_unix "-DGC_STRESS $PARANOID_"
+test_unix "-DCOMPRESSED_CONS $PARANOID"
+test_unix "-DNO_ONERROR $PARANOID"
+test_unix "-DNO_DEBUGGER $PARANOID"
+test_unix "-DNAIVE"